+++ packages/tx-helpers/src/send-btc.ts
@@ -1,10 +1,9 @@
-import { NetworkType } from '@unisat/wallet-types'
+import { NetworkType, ToSignInput } from '@unisat/wallet-types'
 import { ErrorCodes, WalletUtilsError } from './error'
 import { Transaction } from './transaction/transaction'
 import { utxoHelper } from './transaction/utxo'
 import { UnspentOutput } from './types'
 import { bitcoin, UTXO_DUST } from '@unisat/wallet-bitcoin'
-import { ToSignInput } from '@unisat/keyring-service'
 
 export async function sendBTC({
   btcUtxos,