+++ packages/tx-helpers/src/send-runes.ts
@@ -7,8 +7,7 @@ import { Transaction } from './transaction/transaction'
 import { utxoHelper } from './transaction/utxo'
 import { UnspentOutput } from './types'
 import { bitcoin } from '@unisat/wallet-bitcoin'
-import { NetworkType } from '@unisat/wallet-types'
-import { ToSignInput } from '@unisat/keyring-service'
+import { NetworkType, ToSignInput } from '@unisat/wallet-types'
 
 // only one arc20 can be send
 export async function sendRunes({