+++ packages/contact-book/CHANGELOG.md
@@ -1,1 +1,8 @@
 # @unisat/contact-book
+
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @unisat/wallet-types@0.2.0