+++ packages/permission-service/CHANGELOG.md
@@ -1,1 +1,8 @@
 # @unisat/permission-service
+
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @unisat/wallet-types@0.2.0