+++ packages/wallet-bitcoin/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @unisat/wallet-bitcoin
 
+## 0.2.0
+
+### Minor Changes
+
+- reorg
+
+### Patch Changes
+
+- Updated dependencies
+  - @unisat/wallet-types@0.2.0
+
 ## 0.1.13
 
 ### Patch Changes