+++ packages/wallet-types/CHANGELOG.md
@@ -1,1 +1,7 @@
 # @unisat/wallet-types
+
+## 0.2.0
+
+### Minor Changes
+
+- reorg